Fix Videodevil Space Order
When items with non-letter, non-space characters they are displaced
before items that start with a space. This causes videos to be displayed
before Categories, Sort, Next Page and Search icons. Striping these
characters from the left side of the string text fixes this issue. I am
not sure if its necessary to do the same to .curr section